Fixes #19 - env data not generated 'cause of bail
rvm crashes out with a warning if `$GEM_HOME/bin` isn't the first $PATH value which then doesn't allow any environment variables to be created for this script to utilize.  RVM will work here by simply placing the value it expects in the beginning of the PATH environment variable for the temporary environment that this is grabbing data from.  Fixes #19